Added Doxygen Documentation to all .H files for all non getters/setters

+++ Verkeerssimulatie/Baan.cpp
@@ -32,7 +32,7 @@ void Baan::setNaam(const std::string &naam) {
     Baan::fNaam = naam;
 }
 
-const double &Baan::getSnelheidsLimiet() const {
+const int &Baan::getSnelheidsLimiet() const {
     return fSnelheidsLimiet;
 }
